Format code into master

See merge request mareasperez7/horarios!1

+++ apps/docente_area/views.py
@@ -1,3 +1,5 @@
+import re
+
 from django.shortcuts import get_object_or_404
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
@@ -19,7 +21,7 @@ class DocenteAreaConArgumento(APIView):
             serializer = DocenteAreaSerializer(docenteArea)
             return Response(dict(docenteArea=serializer.data))
         except:
-            return Response(dict(Detail="not found"))
+            return Response(dict(detail="not found"))
 
     def put(self, request, pk):
         saved_docenteArea = get_object_or_404(
@@ -64,17 +66,21 @@ class DocenteAreaMixed(APIView):
     permission_classes = (IsAuthenticated,)
 
     def get(self, request, clave, value):
+        if re.search('[a-zA-Z]', value):
+            return Response(dict(detail=f'Error en valor: {value} al buscar {clave.split("_")[0]}'))
         if clave == "docente_id":
             docenteAreas = DocenteArea.objects.filter(da_docente=value)
             if docenteAreas:
                 serializer = DocenteAreaSerializer(docenteAreas, many=True)
                 return Response(dict(docenteArea=serializer.data))
             else:
-                return Response(dict(Detail="not found"))
+                return Response(dict(detail="not found"))
         else:
             return Response(dict(detail="not found"))
 
     def put(self, request, clave, value):
+        if re.search('[a-zA-Z]', value):
+            return Response(dict(detail=f'Error en valor: {value} al buscar {clave.split("_")[0]}'))
         if clave == "docente_id":
             DocenteArea.objects.filter(da_docente=value).delete()
             data = request.data.get("docenteArea")
@@ -92,5 +98,5 @@ class DocenteAreaMixed(APIView):
                         msg = dict(
                             success=f" las Areas del docente '{docenteArea_saved.da_docente}' updated successfully")
         else:
-            msg = dict(Detail="clave invalida")
+            msg = dict(detail="clave invalida")
         return Response(msg)
